Lägga till Calculator Test class

+++ Test_project/src/test1/Calculator.java
@@ -2,7 +2,7 @@ package test1;
 
 public class Calculator {
 	
-		public static int add(int number1, int number2) {
+		public static double add(double number1, double number2) {
 		return number1 + number2;	
 		}
